variable mapping low range decimal

+++ GSC.Data/Dto/LabManagement/LabManagementVariableMappingDto.cs
@@ -32,10 +32,10 @@ namespace GSC.Data.Dto.LabManagement
     {
         public string ProjectDesignVariable { get; set; }
         public string TargetVariable { get; set; }
-        public int? MaleLowRange { get; set; }
-        public int? MaleHighRange { get; set; }
-        public int? FemaleLowRange { get; set; }
-        public int? FemaleHighRange { get; set; }
+        public decimal? MaleLowRange { get; set; }
+        public decimal? MaleHighRange { get; set; }
+        public decimal? FemaleLowRange { get; set; }
+        public decimal? FemaleHighRange { get; set; }
         public string Unit { get; set; }
         public string AuditReason { get; set; }
         public string ReasonOth { get; set; }
